[Backtracer][DWARF] Fix SLEB128 sign extension code.

+++ stdlib/public/RuntimeModule/Dwarf.swift
@@ -346,7 +346,7 @@ extension ImageSource {
     }
 
     if shift < 64 && sign != 0 {
-      value |= -(1 << shift)
+      value |= Int64(-1) << shift
     }
 
     return (addr, value)
